Validate parameters of push apis, and add small tests

+++ src/ops/PushOps.ts
@@ -3,7 +3,7 @@ import {Promise} from 'es6-promise';
 import request from './Request';
 import {App} from '../App';
 import {APIAuthor} from '../APIAuthor';
-import {HttpRequestError} from '../ThingIFError'
+import {ThingIFError, HttpRequestError, Errors} from '../ThingIFError'
 import MqttInstallationResult from '../MqttInstallationResult'
 import {Response} from './Response'
 import BaseOp from './BaseOp'
@@ -40,6 +40,14 @@ export default class CommandOps extends BaseOp {
         development: boolean
         ): Promise<string> {
         return new Promise<string>((resolve, reject) => {
+            if (!installationRegistrationID){
+                reject(new ThingIFError(Errors.ArgumentError, "installationRegistrationID is null or empty"));
+                return;
+            }
+            if (!development){
+                reject(new ThingIFError(Errors.ArgumentError, "development is null"));
+                return;
+            }
             let requestBody = {
                 installationRegistrationID: installationRegistrationID,
                 deviceType: "ANDROID",
@@ -56,25 +64,22 @@ export default class CommandOps extends BaseOp {
     }
 
     installMqtt(development: boolean): Promise<MqttInstallationResult> {
-
         return new Promise<MqttInstallationResult>((resolve, reject) => {
+            if (!development){
+                reject(new ThingIFError(Errors.ArgumentError, "development is null"));
+                return;
+            }
+
             let requestBody = {
                 deviceType: "MQTT",
                 development: development
             };
             this.installPush(requestBody).then((res)=>{
                 let body = res.body;
-                let installationID = (<any>body).installationID;
-                let installationRegistrationID = (<any>body).installationRegistrationID;
-                if(!!installationID && !! installationRegistrationID) {
-                    let result = new MqttInstallationResult(installationID, installationRegistrationID);
+                    let result = new MqttInstallationResult(
+                        (<any>body).installationID,
+                        (<any>body).installationRegistrationID);
                     resolve(result);
-                }else{
-                    let err = new Error();
-                    err.name = "InvalidResponse"
-                    err.message = "No installationID or installationRegistrationID in response: "+ JSON.stringify(res);
-                    reject(err);
-                }
             }).catch((err)=>{
                 reject(err);
             })
@@ -82,9 +87,12 @@ export default class CommandOps extends BaseOp {
     }
 
     uninstall(installationID: string): Promise<void> {
-
         let url = `${this.au.app.getKiiCloudBaseUrl()}/installations/${installationID}`;
         return new Promise<void>((resolve, reject) => {
+            if (!installationID){
+                reject(new ThingIFError(Errors.ArgumentError, "installationID is null or empty"));
+                return;
+            }
             request({
                 method: "DELETE",
                 headers: this.getHeaders(),
